requirements.txt
google-adk
fastapi
jinja2